requirements.txt
python-dotenv
requests
solana
solders